fix(Parser): Converting short flag name (u8) to []const u8 by using short lived reference
When parseShortArg sets the err_ctx.proivded_arg by calling
err_ctx.setProvidedArg before setting it first it converts flag (u8) (which is being
returned by short_flag.next()) to []const u8 which then passed to
err_ctx.setProvidedArg then later when we try to print it outputs empty
string because the []const u8 was holding the stack reference which
used to be destory when it leaves parseShortArg. This patch fixes this
issue by making ShortFlag.next to return ?*const u8 instead.