fix(Parser): Use a pointer capture instead of value.
Before this patch when `parseCommandArgument` used to call `consumeArgValue`
by passing address of `arg` it was actually passing the stack address
instead of `arg.items`. So when `consumeArgValue` sets the
`ErrorContext.arg` it was actually setting with the address of stack
which would later causes `ErrorContext.logError` to print weired output.